Caster attack now damages player

+++ Assets/Scripts/PlayerController.cs
@@ -33,8 +33,8 @@ public class PlayerController : MonoBehaviour
     [SerializeField] private GameObject areaSpell;
 
     [SerializeField] private int lives;
-    [SerializeField] private int maxHealth = 100;
-    private int currentHealth;
+    [SerializeField] private float maxHealth = 100;
+    private float currentHealth;
 
     [SerializeField] private Transform spawnLocation;
 
@@ -104,7 +104,7 @@ public class PlayerController : MonoBehaviour
     }
 
 
-    public int GetHealth() {
+    public float GetHealth() {
         return currentHealth;
     }
 
@@ -186,7 +186,7 @@ public class PlayerController : MonoBehaviour
         
     }
 
-    public void TakeDamage(int damageToTake)
+    public void TakeDamage(float damageToTake)
     {
 
         currentHealth -= damageToTake;
